requirements.txt
certifi==2020.12.5
cffi==1.14.0
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.1.3
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
nbconvert==5.6.1
olefile==0.46
Pillow 
protobuf==3.15.6
pycparser
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
six
tensorboardX==1.6
pytorch==1.7.1
torchgeometry==0.1.2
torchvision==0.2.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
